--story=886002359 云桌面实例管理

+++ src/backend/ci/ext/tencent/remotedev/api-remotedev-tencent/src/main/kotlin/com/tencent/devops/remotedev/pojo/ProjectWorkspaceCreate.kt
@@ -33,7 +33,7 @@ import io.swagger.annotations.ApiModelProperty
 @ApiModel("工作空间信息-创建")
 data class ProjectWorkspaceCreate(
     @ApiModelProperty("windows 配置id")
-    val windowsResourceConfigId: Long,
+    val windowsResourceConfigId: Int,
     @ApiModelProperty("基础镜像Id")
     val baseImageId: Int,
     @ApiModelProperty("创建实例的数量")
